requirements.txt
fastapi
uvicorn
joblib
pandas
numpy
scikit-learn
requests
feedparser
beautifulsoup4
googlesearch-python
kaggle
huggingface_hub
python-nmap
scapy
